chore: move dynamicValue to new file
