+++ src/blueprint/mod.rs
@@ -1,6 +1,7 @@
 mod blueprint;
 mod compress;
 mod definitions;
+mod dynamic_value;
 mod from_config;
 mod into_schema;
 mod links;
@@ -11,8 +12,10 @@ mod schema;
 mod server;
 mod timeout;
 mod upstream;
+
 pub use blueprint::*;
 pub use definitions::*;
+pub use dynamic_value::*;
 pub use from_config::*;
 pub use links::*;
 pub use operation::*;
@@ -71,8 +74,8 @@ impl TypeLike for Arg {
 }
 
 pub(crate) fn to_type<T>(field: &T, override_non_null: Option<bool>) -> Type
-where
-    T: TypeLike,
+    where
+        T: TypeLike,
 {
     let name = field.name();
     let list = field.list();
