+++ src/blueprint/operators/const_field.rs
@@ -1,6 +1,5 @@
-use async_graphql_value::{ConstValue, Name};
+use async_graphql_value::ConstValue;
 use indexmap::IndexMap;
-use serde_json::Value;
 
 use crate::blueprint::*;
 use crate::config;
@@ -43,56 +42,6 @@ impl MustacheOrValue {
         match self {
             MustacheOrValue::Mustache(m) => m.is_const(),
             _ => true,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum DynamicValue {
-    Value(serde_json::Value),
-    Mustache(Mustache),
-    MustacheObject(IndexMap<Name, MustacheOrValue>),
-    MustacheArray(Vec<MustacheOrValue>),
-}
-
-impl TryFrom<&serde_json::Value> for DynamicValue {
-    type Error = anyhow::Error;
-
-    fn try_from(value: &Value) -> Result<Self, Self::Error> {
-        match value {
-            Value::Object(obj) => {
-                let mut out = IndexMap::new();
-                for (k, v) in obj {
-                    let m = Mustache::parse(v.to_string().as_str())?;
-                    if m.is_const() {
-                        out.insert(Name::new(k), MustacheOrValue::Value(v.clone()));
-                    } else {
-                        out.insert(Name::new(k), MustacheOrValue::Mustache(m));
-                    }
-                }
-                Ok(DynamicValue::MustacheObject(out))
-            }
-            Value::Array(arr) => {
-                let mut out = Vec::new();
-                for v in arr {
-                    let m = Mustache::parse(v.to_string().as_str())?;
-                    if m.is_const() {
-                        out.push(MustacheOrValue::Value(v.clone()));
-                    } else {
-                        out.push(MustacheOrValue::Mustache(m));
-                    }
-                }
-                Ok(DynamicValue::MustacheArray(out))
-            }
-            Value::String(s) => {
-                let m = Mustache::parse(s.as_str())?;
-                if m.is_const() {
-                    Ok(DynamicValue::Value(value.clone()))
-                } else {
-                    Ok(DynamicValue::Mustache(m))
-                }
-            }
-            _ => Ok(DynamicValue::Value(value.clone())),
         }
     }
 }
@@ -145,8 +94,7 @@ pub fn compile_const(inputs: CompileConst) -> Valid<Expression, String> {
     }
 }
 
-pub fn update_const_field<'a>(
-) -> TryFold<'a, (&'a ConfigModule, &'a Field, &'a config::Type, &'a str), FieldDefinition, String>
+pub fn update_const_field<'a>() -> TryFold<'a, (&'a ConfigModule, &'a Field, &'a config::Type, &'a str), FieldDefinition, String>
 {
     TryFold::<(&ConfigModule, &Field, &config::Type, &str), FieldDefinition, String>::new(
         |(config_module, field, _, _), b_field| {
@@ -158,10 +106,10 @@ pub fn update_const_field<'a>(
                 DynamicValue::try_from(&const_field.data.clone())
                     .map_err(|e| ValidationError::new(e.to_string())),
             )
-            .and_then(|value| {
-                compile_const(CompileConst { config_module, field, value: &value, validate: true })
-                    .map(|resolver| b_field.resolver(Some(resolver)))
-            })
+                .and_then(|value| {
+                    compile_const(CompileConst { config_module, field, value: &value, validate: true })
+                        .map(|resolver| b_field.resolver(Some(resolver)))
+                })
         },
     )
 }
